Refactored a little bit more. Temporary change to trigger nighttly from this branch
