+++ e2e-tests/src/synthetic_network.rs
@@ -4,10 +4,15 @@ use aleph_client::{
     SignedConnection,
 };
 use anyhow::anyhow;
+use futures::future::join_all;
 use log::info;
 use synthetic_link::SyntheticNetworkClient;
 
+use crate::config::Config;
+
 pub type Milliseconds = u64;
+
+const OUT_LATENCY: u64 = 200;
 
 pub async fn set_out_latency(milliseconds: Milliseconds, synthetic_url: String) {
     info!(
@@ -48,5 +53,36 @@ pub async fn wait_for_further_finalized_blocks(
     connection
         .wait_for_block(|n| n > block_number_to_wait, BlockStatus::Finalized)
         .await;
+    Ok(())
+}
+
+pub async fn test_latency_template_test(
+    config: &Config,
+    validator_count: usize,
+) -> anyhow::Result<()> {
+    let out_latency = config.test_case_params.out_latency.unwrap_or(OUT_LATENCY);
+    let blocks_to_wait: u32 = 30;
+
+    let connections = config.create_signed_connections().await;
+    join_all(
+        connections
+            .iter()
+            .map(|connection| wait_for_further_finalized_blocks(connection, blocks_to_wait)),
+    )
+    .await;
+    join_all(
+        config
+            .synthetic_network_urls()
+            .into_iter()
+            .take(validator_count)
+            .map(|synthetic_url| set_out_latency(out_latency, synthetic_url)),
+    )
+    .await;
+    join_all(
+        connections
+            .iter()
+            .map(|connection| wait_for_further_finalized_blocks(connection, blocks_to_wait * 2)),
+    )
+    .await;
     Ok(())
 }
